requirements.txt
openai==1.12.0
pandas==2.2.0
plotly==5.18.0
python-dotenv==1.0.1
pytest==8.0.0
streamlit==1.31.0
watchdog==3.0.0